Fix operator== for input iterators
For forward iterators, the position of base() is at the start of the
current code point. For input iterators, it's immediately after the
end of the current code point.

For forward iterators, checking whether we're at the end is simple,
because we can just check whether base() is equal to end().

For input iterators, if base() is equal to end(), we may still be in
the middle of the code point immediately preceding end().

The old way of discriminating whether we were past the last code point
was checking if buf_index_ was equal to buf_.size(). However, this
isn't viable because we might want to construct a transcoding iterator
at the end by passing in an input iterator at the end of the sequence
and then we have no information about what buf_.size() would be if we
had processed the last code point.

So we instead need some kind of at-the-end sentinel in the input case,
which we achieve by setting buf_index_ to -1.